Check for the status of a submitted search on NCBI's CD-SEARCH.

Check on the status of a submitted search associated with the given cdsid (@seealso cdssubmit). The values for all the parameters below are inherited from cdsearchr (@seealso cdsearchr() for some details). This function is also responsible for interpreting (and printing to the terminal) the CD-SEARCH-specific error-codes.

cdscheck(cdsid_check = NULL, tdata_check = NULL, cdsurl_check = NULL,
maxreps = NULL, sleep_duration = NULL)

Arguments

cdsid_check

(character string, mandatory) the cdsid returned by the NCBI CD-SEARCH server upon successful submission of a search request.

tdata_check

(character string, mandatory) the format in which the target data should be retrieved.

cdsurl_check

(character string, mandatory) URL at which the CD-SEARCH server can be accessed.

maxreps

(integer, mandatory) maximum number of attempts to be made to query the server on the status of the search before giving up.

sleep_duration

(integer, mandatory) maximum duration in seconds between successive queries to the server.

Value

1 if the search terminated successfully, otherwise 0.