Wrapper around base::list.files() that adds an option to exclude directories automatically.

list.files returns all objects in the search path including directories. list_files is a wrapper that adds an option to exclude directories. In specific, it allows the user to override the default behavior of list.files() including subdirectory names when recursive is set to FALSE. This is controlled by the incl_dirs argument which filters our directories when set to TRUE.

list_files(path = ".", pattern = NULL, all.files = FALSE, full.names = TRUE,
recursive = FALSE, ignore.case = FALSE, include.dirs = FALSE, no.. = FALSE, incl_dirs = FALSE)

Arguments

path

a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored. Elements with a marked encoding will be converted to the native encoding (and if that fails, considered non-existent).

pattern

an optional regular expression. Only file names which match the regular expression will be returned.

all.files

a logical value. If FALSE, only the names of visible files are returned (following Unix-style visibility, that is files whose name does not start with a dot). If TRUE, all file names will be returned.

full.names

a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.

recursive

logical. Should the listing recurse into directories?

ignore.case

logical. Should pattern-matching be case-insensitive?

include.dirs

logical. Should subdirectory names be included in recursive listings? (They always are in non-recursive ones).

no..

logical. Should both "." and ".." be excluded also from non-recursive listings?

incl_dirs

(binary, optional) set to FALSE to remove all directories from the search results irrespective of whether or not recursive=TRUE.

See also

Examples

if (FALSE) {
#Input data
inpath <- dirname(system.file("extdata", "cdsearchr_testdata.fasta",
                              package = "seqvisr", mustWork = TRUE))
#Listing files
list_files(inpath)
}