Wrapper script around tsvtogginp() to perform conversion of a data.frame() of multiple sequences and their feature annotations.

Wrapper script around tsvtogginp() to perform conversion of a data.frame() of multiple sequences and their feature annotations.

tsvtogginp_multi(inpdf = NULL)

Arguments

inpdf

(character, mandatory) the name of the input data.frame().

Value

a data.frame() with one row per residue indicating the sequence name, position (indicated by that row), the description associated with that position (e.g., "ABC domain", or NA for no description), a column indicating which feature the position belongs to (e.g., say there are two domains named "ABC", this numerical column is a way to distinguish between them), a column indicating the layer (offset, base, feature; see ?pdomvisr()), label for the particular feature (assigned to one position within the feature; see ?pdomvisr()), and a column indicating the original positions of the rows (e.g., when a sequence offset has been applied; see ?pdomvisr()).

Details

Basically applies tsvtogginp() on multiple sequences worth of rows using dplyr::group_map(). tsvtogginp_multi() is invoked by pdomvisr().

See also

Examples

if (FALSE) {
#Input data
inpath <- system.file("extdata", "pdomvisr_testdata.tsv", package = "seqvisr", mustWork = TRUE)
inpdat <- read.table(inpath, header = TRUE)

#Example run
tsvtogginp_multi(inpdat)
}